Completed the T3.Account.Api project

+++ src/server/T3.Account.Api/Controllers/AccountController.cs
@@ -1,4 +1,6 @@
+using System.Net;
 using Microsoft.AspNetCore.Mvc;
+using T3.Account.Api.Errors;
 using T3.Account.Api.Models;
 using T3.Account.Api.Services;
 
@@ -8,32 +10,44 @@ namespace T3.Account.Api.Controllers;
 [Route("[controller]")]
 public class AccountController : ControllerBase
 {
-    private readonly ILogger<AccountController> _logger;
     private readonly IAccountCreateService _accountCreateService;
     private readonly IAccountChangePasswordService _accountChangePasswordService;
     private readonly IAccountDeleteService _accountDeleteService;
+    private readonly IAccountLoginService _accountLoginService;
 
-    public AccountController(ILogger<AccountController> logger, IAccountCreateService accountCreateService, IAccountChangePasswordService accountChangePasswordService, IAccountDeleteService accountDeleteService)
+    public AccountController(IAccountCreateService accountCreateService, IAccountChangePasswordService accountChangePasswordService, IAccountDeleteService accountDeleteService, IAccountLoginService accountLoginService)
     {
-        _logger = logger;
         _accountCreateService = accountCreateService;
         _accountChangePasswordService = accountChangePasswordService;
         _accountDeleteService = accountDeleteService;
+        _accountLoginService = accountLoginService;
     }
 
     [HttpPost]
     public async Task<CreateAccountResponse> Create(CreateAccountRequest request) => await _accountCreateService.Create(request);
     
     [HttpPost("change-password")]
-    public async Task<ChangePasswordResponse> Change(ChangePasswordRequest request) => await _accountChangePasswordService.ChangePassword(request);
+    public async Task<ChangePasswordResponse> ChangePassword(ChangePasswordRequest request) => await _accountChangePasswordService.ChangePassword(request);
 
     [HttpDelete("{accountId:guid}")]
     public async Task Delete(Guid accountId) 
     {
-        // TODO: Add check if user is authorized to delete this account
+        if(User.Identity?.Name != accountId.ToString())
+            throw new HttpRequestException("You are not authorized to delete this account", null, HttpStatusCode.Forbidden);
+
         await _accountDeleteService.Delete(new DeleteAccountRequest(accountId));
     }
 
     [HttpPost("login")]
-    public async Task<LoginResponse> Login(LoginRequest request) => throw new NotImplementedException();
+    public async Task<LoginResponse> Login(LoginRequest request)
+    {
+        try
+        {
+            return await _accountLoginService.Login(request);
+        }
+        catch (LoginException ex)
+        {
+            throw new HttpRequestException("Username or password is incorrect", ex, HttpStatusCode.BadRequest);
+        }
+    }
 }
